users-Firebase ergänzt zur Levelverwaltung Leverberechnung und Punktevergabe ergänzt Darstellung des Levels hinzugefügt

+++ app/src/main/java/com/thepublichistorian/heritagehunter/models/Place.kt
@@ -6,7 +6,7 @@ data class Place(
     val name: String = "",
     val latitude: Double = 0.0,
     val longitude: Double = 0.0,
-    val counter: Int = 0,
+    val counter: Long = 0,
     val visitors: List<String> = listOf(),
     val category: String = ""
 )
